Multiple answers polls
Please delete anketa.js
